# fix benchmark output